add HuberCostFun to LinearRegression.scala
